+++ src/lib/ethService.ts
@@ -1,8 +1,9 @@
+import { createAlchemyWeb3 } from "@alch/alchemy-web3";
 import { ethers } from "ethers";
 import { generateNonce, SiweMessage } from "siwe";
 
 // TODO: Move to a secure location.
-const API_URL = "https://eth-goerli.alchemyapi.io/v2/_ArbR3W9Ttz3Cx1Ofa_vVhcIxzxbx7tC";
+const API_URL = "https://eth-mainnet.alchemyapi.io/v2/_ArbR3W9Ttz3Cx1Ofa_vVhcIxzxbx7tC";
 const PRIVATE_KEY = "_ArbR3W9Ttz3Cx1Ofa_vVhcIxzxbx7tC";
 
 export { ethService };
@@ -13,6 +14,7 @@ class EthService {
   address = "";
   message = "";
   signature = "";
+  web3 = createAlchemyWeb3(API_URL);
 
   constructor() {
     console.log({ API_URL, PRIVATE_KEY });
@@ -29,8 +31,6 @@ class EthService {
   requestAccounts = async () => {
     try {
       const out = await this.provider.send("eth_requestAccounts", []);
-      console.log({ out });
-
       const address = out[0];
       return address;
     } catch (e) {
@@ -44,23 +44,13 @@ class EthService {
     const address = await this.requestAccounts();
     this.address = address;
 
-    // const name = await this.provider.lookupAddress(address);
-    // console.log(name);
-
-    // const balance = await this.provider.getBalance(address);
-
-    // balance is a BigNumber (in wei); format is as a sting (in ether)
-    // var etherString = ethers.utils.formatEther(balance);
-    // console.log("Balance: " + etherString);
-
-    // const transactionCount = await this.provider.getTransactionCount(address);
-    // console.log("Total Transactions Ever Send: " + transactionCount);
+    console.log({ address: this.address });
   };
 
   signIn = async () => {
     this.message = await this.createSiweMessage(await this.signer.getAddress(), "Sign in with Ethereum to the app.");
     this.signature = await this.signer.signMessage(this.message);
-    console.log({ message: this.message, signature: this.signature });
+    console.log({ message: this.message, address: this.address, signature: this.signature });
   };
 
   signMessage = async (statement: string) => {
@@ -72,6 +62,32 @@ class EthService {
       console.error(e);
     }
   };
+
+  getBlock = async () => {
+    const blockNumber = await this.web3.eth.getBlockNumber();
+    console.log("The latest block number is " + blockNumber);
+  };
+
+  getBalances = async () => {
+    const result = await this.web3.alchemy.getTokenBalances(this.address);
+    return result.tokenBalances;
+  };
+
+  getTokenBalanceMeta = async (addr: string) => {
+    return await this.web3.alchemy.getTokenMetadata(addr);
+  };
 }
 
 const ethService = new EthService();
+
+// const name = await this.provider.lookupAddress(address);
+// console.log(name);
+
+// const balance = await this.provider.getBalance(address);
+
+// balance is a BigNumber (in wei); format is as a sting (in ether)
+// var etherString = ethers.utils.formatEther(balance);
+// console.log("Balance: " + etherString);
+
+// const transactionCount = await this.provider.getTransactionCount(address);
+// console.log("Total Transactions Ever Send: " + transactionCount);